+++ packages/docstage/src/docstage/app_keys.py
@@ -1,9 +1,7 @@
 """Application keys for type-safe app configuration access."""
 
 from aiohttp import web
-from docstage_core import PageRenderer
-
-from docstage.core.site import SiteLoader
+from docstage_core import PageRenderer, SiteLoader
 
 renderer_key = web.AppKey("renderer", PageRenderer)
 site_loader_key = web.AppKey("site_loader", SiteLoader)